requirements.txt
uvicorn==0.22.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.15
psycopg2-binary==2.9.6
alembic==1.11.1
minio==7.1.15
redis==4.5.5
PyPDF2==3.0.1
python-docx==0.8.11
fastapi==0.109.0
pydantic==2.6.0
celery==5.3.6
asyncpg>=0.29.0
pydantic-settings>=2.0.0